Patient is now safe from SQL Injections
Insert and Update are the only two that need to be secure until we get
an Advanced Search working.  First create your routines in MYSQL
Workbench(Insert and Update), then change your insert and update so that
they use the same function calls as Patient.  Database.cs and
Mysqlwrapper should do the rest.