Notifications Functionality Added
A text box has been added to the bottom of the Main Tab View, which
polls the database every 5 seconds for new information. If you edit this
box and then remove keyboard focus (Tab key), it will update the
database.

Tweaking these values might be needed in the future for better user
experience. But it works as described above.